# dequantize + matrix multiplication CUDA kernels