I implemented CUDA kernels that do dequantization and matrix matrix multiplication in one step. This eliminates the need for temporary buffers to hold the dequantized f32 matrix given to cuBLAS. For 33b this saves at least ~600 MiB for -ngl <= 60, and at least ~1400 MiB for -ngl >= 61. As a result a few more layers can be offloaded. For this table pp == prompt processing, tg128 == generation of 128 tokens with an empty prompt: 

| GPU      | Model      | Test  | Max. -ngl master | t/s master | Max -ngl PR | t/s PR |
|----------|------------|-------|------------------|------------|-------------|--------|
| RTX 3090 | 33b q5_1   | pp    |               53 |        131 |          57 |     61 |
| RTX 3090 | 33b q5_1   | tg128 |               55 |       7.86 |          58 |   9.90 |
| RTX 3090 | 33b q5_k_m | pp    |               55 |     135.19 |          61 |  53.73 |
| RTX 3090 | 33b q5_k_m | tg128 |               58 |      11.39 |          61 |  14.77 |

Unfortunately my matrix multiplication kernel is not very good and significantly slower than cuBLAS for prompt processing. This is particularly noticeable when there is plenty of VRAM anyways:

| GPU      | Model     | Test | t/s master | t/s PR |
|----------|-----------|------|------------|--------|
| RTX 3090 | 7b f16    | pp   |       1023 |    516 |
| RTX 3090 | 7b q4_0   | pp   |       1048 |    477 |
| RTX 3090 | 7b q4_1   | pp   |       1042 |    443 |
| RTX 3090 | 7b q5_0   | pp   |       1032 |    383 |
| RTX 3090 | 7b q5_1   | pp   |       1032 |    359 |
| RTX 3090 | 7b q8_0   | pp   |       1029 |    476 |
| RTX 3090 | 7b q2_K   | pp   |       1028 |    371 |
| RTX 3090 | 7b q3_K_m | pp   |       1009 |    310 |
| RTX 3090 | 7b q4_K_m | pp   |       1010 |    352 |
| RTX 3090 | 7b q5_K_m | pp   |       1040 |    308 |
| RTX 3090 | 7b q6_K   | pp   |       1033 |    372 |

By default cuBLAS is still in use. The new dequantization + matrix multiplication kernels can be used by setting the compile option `LLAMA_CUDA_DMM`.

The implementation works by adding a template that takes a dequantization method. I revised the dequantization to require only the index of the data value to be dequantized. I plan to eventually apply this dequantization scheme for the other templates as well since it is simpler.

I don't understand why the performance of my matrix multiplication kernel is so bad. I tried several other variants but they all had even worse performance. I would very much welcome it if someone were to write a better one.