@JohannesGaessler 2023-06-28
While I was working on this the PR for k-quant super block sizes of 64 https://github.com/ggerganov/llama.cpp/pull/2001 was merged. This option is not supported by my implementation.