@cmp-nct 2023-06-28
Quite good results, not a small feat despite not matching the speed.
Painful to see it's up to 3x faster in cuBLAS but anyway it's one potential (chunky) dependency less without cuBLAS.