@Green-Sky 2023-06-28
Making more room in vram matters alot on constrained system.