@ggerganov 2023-06-29
@JohannesGaessler 

I think you might be heading in a wrong direction - let me try to explain

Before starting work on the GPU code, we were focused on implementing a highly-optimized version for CPU-only inference. The key insight was to start using quantized data during matrix multiplication in order to reduce the memory bandwidth and improve the inference performance. To do that, we developed optimized SIMD "kernels" for computing the dot product **directly working with the quantized integer data**. I.e. it is essential that we **do not dequantize** back to F16 or F32, but instead **quantize `src1` to `Q8`** and perform the `QX - Q8` dot product using integer intrinsic. This is much faster than anything else since it minimizes the required memory throughput and it outperforms all CPU BLAS implementations, simply because all of them require 32-bit FP precision. We have also extensively studied the precision of the quantized dot products and we showed that it is good enough (https://github.com/ggerganov/llama.cpp/pull/951)

This was the main reason during your first CUDA PR to ask you to investigate this same approach for the GPU:

https://github.com/ggerganov/llama.cpp/pull/1375#issuecomment-1545813471

Although you showed that this approach was not efficient on older GPUs, I think we should revisit this and put some more effort in making it work. The expectation is that using it for `matrix x matrix` multiplication will be much faster than F16 / F32 cuBLAS due to the same reasons that we found and confirmed during the CPU-only development.

The existing `QX - Q8` dot products are a very good start for `matrix - vector` multiplications, but are not the best for `matrix x matrix` multiplications. At the very least, we are missing the block-tiling optimization which should allow us to utilize the L2 cache much better.

The goal of https://github.com/ggerganov/ggml/issues/293 will be to add exactly this type of optimization (and potentially other). We will start with the CPU implementation and then I plan to do the same for Metal. I'm pretty confident that this is the correct approach to take and will recommend to do the same for CUDA and OpenCL.