@slaren 2023-06-29
I think that dequantizing to f16 or f32 is going to be necessary to use the tensor cores, so I think this approach has merit. Dequantizing to SRAM should still be much faster than the previous dequantization to DRAM, and if the bottleneck is the compute (and it should be), then tensor cores should improve performance significantly, possibly more than integer math.