@JohannesGaessler 2023-06-29
Regarding floats vs ints: First of all I think it should be noted that CPUs and GPUs have very different performance characteristics. In particular:

1. Floating point arithmetic is comparatively much faster on GPUs than it is on CPUs.
2. (NVIDIA) GPUs are able to load 32-128 contiguous bytes at once. So if a GPU loads two q8_0 blocks it needs at least 4 memory accesses (because the current blocks have bad memory alignment) to load the quants and the scales compared to 2 (larger) memory accesses to load two contiguous segments of 32 f16/f32 values.
4. Given the correct implementation matrix matrix multiplication should **not** be I/O bound. If there is a benefit from using quantized data it will come from being able to use less shared memory for the tiles. In particular, my implementation previously did f32 + f32 -> f32 matrix matrix multiplication without on-the-fly dequantization and the performance was essentially the same (but I think an implementation utilizing larger tiles would still be beneficial in terms of reducing I/O).
5. Per-byte integer intrinsics, particularly those using tensor cores, are **not** available on all GPUs (or they have bad performance). So we will need an f32 implementation anyways and it serves as a baseline to compare the efficiency of the matrix multiplication kernel itself to a highly optimized implementation, i.e. cuBLAS. 
6. On GPUs you cannot save clock cycles by doing the integer dot product first and then scaling afterwards (when not using integer intrinsics). On a CPU you would have to do 96 multiplications if you dequantize first and then do the dot product compared to 34 multiplications if you first do the dot product and then dequantize. On an NVIDIA GPU both take 4 multiplications because you can do 32 multiplications in parallel.
7. Tensor cores can use f32, f16, and int8 as input data so dequantization is not a prerequisite for using tensor cores. Sub-byte integer intrinsics currently do not have a stable API.

I think currently the biggest improvement for quantized CUDA performance to be had will be to rearrange the data and to utilize shared memory. For CUDA to efficiently load data it needs to be aligned to 32 bytes. However, the current blocks have sizes of e.g. 18 bytes. So what should be done I think is to rearrange the data when it's transferred into VRAM so that all quants come first followed by all scales. Then, when the data is used the quants are memory aligned and the scales can be read into shared memory as blocks (since they take up comparatively few bytes). Of course this is not mutually exclusive with using quantized dot products though.