@slaren 2023-06-29
>Tensor cores can use f32, f16, and int8 as input data so dequantization is not a prerequisite for using tensor cores. Sub-byte integer intrinsics currently do not have a stable API.

Right, but the intermediate results need higher precision than int4 or even int8. On the CPU we usually do the dot products in int16 before scaling to float. I am not sure if that's possible to do with tensor cores, but I don't think it is.  We could do the same in the GPU, just without tensor cores, and it would likely be faster than the current implementation, but I am not sure that it would be faster than converting to f16 and using the tensor cores.

The int4 ops have been deprecated in Hopper, so I don't expect that they will ever have a stable API.