@JohannesGaessler 2023-06-29
>Right, but the intermediate results need higher precision than int4 or even int8.

According to the [documentation](https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#element-types-and-matrix-sizes) the accumulator for int8 tensor cores is int32.