@cmp-nct 2023-06-29
> > Right, but the intermediate results need higher precision than int4 or even int8.
> 
> According to the [documentation](https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#element-types-and-matrix-sizes) the accumulator for int8 tensor cores is int32.

Good to know. Though the 2nd tensor is usually so small that on-the-fly conversion is always possible. In my fblas-16bit I'm just down-converting the 2nd matrix inside the function, I didn't notice a performance difference.

I wonder if Nvidia will actually release the fp8 functions to consumer hardware, in the past year I had the impression that they actively work against the use of their gaming-GPU line for ML applications. Reducing VRAM, restricting driver licensing, removing SLI. Their datacenter line brings in 2 times more revenue as they just scale the price of large VRAM models up 10 fold.